requirements.txt
streamlit>=1.35.0
pdfplumber>=0.11.0
pandas>=2.0.0

# LLM providers — instalar según el proveedor que uses
openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.7.0
mistralai>=1.0.0
